next/requirements.txt
Flask==3.0.3
langchain==0.3.0
langchain-core==0.3.5
langchain-text-splitters==0.3.0
cerebras_cloud_sdk==1.3.0
python-dotenv==1.0.1